removed static props - added serverside props
