+++ pages/allproducts/index.tsx
@@ -1,12 +1,14 @@
 import { SetStateAction, useState, useEffect } from 'react'
 import Head from 'next/head'
-import type { InferGetStaticPropsType, GetStaticProps } from 'next'
+import type { InferGetServerSidePropsType } from 'next'
+//import type { InferGetStaticPropsType, GetStaticProps } from 'next'
 import Layout from '../../components/Layout'
 import AllProductBoxes from '../../components/allproducts/allProductBoxes'
 import AccordianMenu from '../../components/accordianMenu'
 import SearchBar from '../../components/searchBar/SearchBar'
 import SearchSelect from '../../components/searchSelect/SearchSelect'
 import 'bootstrap/dist/css/bootstrap.min.css'
+import clientPromise from '../../lib/mongodb'
 
 import {FilteredData} from "../../types/index"
 
@@ -17,7 +19,7 @@ type AllData = {
   selectresults: [];
 }
 
-export default function Allproducts(props: InferGetStaticPropsType<typeof getStaticProps>) {
+export default function Allproducts(props: InferGetServerSidePropsType<typeof getServerSideProps>) {
 
   const {product, accordian,searchresults,selectresults } :AllData = props.allData;
 
@@ -110,18 +112,43 @@ export default function Allproducts(props: InferGetStaticPropsType<typeof getSta
   )
 }
 
-export const getStaticProps: GetStaticProps<{
-  allData: AllData
-}> = async () => {
+export const getServerSideProps = async () => {
   // Call an external API endpoint to get posts
-  const res = await fetch('/api/alldata')
-  const allData = await res.json()
+  // const res = await fetch('http://localhost:3000/api/alldata')
+  // const allData = await res.json()
+  try {
+    //await clientPromise
+    const client = await clientPromise
+    const db = client.db('shoestore')
+
+    const results = await db.collection('products').find({}).toArray()
+    const resultstwo = await db.collection('accordianData').find({}).toArray()
+    const resultsthree = await db.collection('selectBarData').find({}).toArray()
+    const resultsfour = await db.collection('searchBarData').find({}).toArray()
+
+    const product = JSON.parse(JSON.stringify(results))
+    const accordian = JSON.parse(JSON.stringify(resultstwo))
+    const searchresults = JSON.parse(JSON.stringify(resultsfour))
+    const selectresults = JSON.parse(JSON.stringify(resultsthree))
+
+    return {
+      props:  {
+        allData :{product,accordian,searchresults,selectresults},
+        revalidate: 10,
+      }
+    }
+    // res.status(200).send({product,accordian,searchresults,selectresults});
+
+  } catch (e) {
+    console.error(e)
+  }
+
   // By returning { props: { posts } }, the Blog component
   // will receive `posts` as a prop at build time
-  return {
-    props:  {
-      allData,
-      revalidate: 10,
-    }
-  }
+  // return {
+  //   props:  {
+  //     allData,
+  //     revalidate: 10,
+  //   }
+  // }
 }
