+++ pages/specials/index.tsx
@@ -1,5 +1,6 @@
 import { SetStateAction, useState, useEffect } from 'react'
-import type { InferGetStaticPropsType, GetStaticProps } from 'next'
+import type { InferGetServerSidePropsType } from 'next'
+//import type { InferGetStaticPropsType, GetStaticProps } from 'next'
 import SpecialsProductBoxes from '../../components/specials/specialsProductBoxes'
 import AccordianMenu from '../../components/accordianMenu'
 import SearchBar from '../../components/searchBar/SearchBar'
@@ -7,10 +8,10 @@ import SearchSelect from '../../components/searchSelect/SearchSelect'
 import Head from 'next/head'
 import Layout from '../../components/Layout'
 import 'bootstrap/dist/css/bootstrap.min.css'
-
+import clientPromise from '../../lib/mongodb'
 import {FilteredData , AllData} from "../../types/index"
 
-export default function Specials(props: InferGetStaticPropsType<typeof getStaticProps> ) {
+export default function Specials(props: InferGetServerSidePropsType<typeof getServerSideProps> ) {
 
   const {product, accordian,searchresults,selectresults } : AllData = props.allData;
 
@@ -106,18 +107,42 @@ export default function Specials(props: InferGetStaticPropsType<typeof getStatic
   )
 }
 
-export const getStaticProps: GetStaticProps<{
-  allData: AllData
-}> = async () => {
+export const getServerSideProps = async () => {
   // Call an external API endpoint to get posts
-  const res = await fetch('/api/alldata')
-  const allData = await res.json()
+  //const res = await fetch('http://localhost:3000/api/alldata')
+  //const allData = await res.json()
   // By returning { props: { posts } }, the Blog component
   // will receive `posts` as a prop at build time
-  return {
-    props:  {
-      allData,
-      revalidate: 10,
+  try {
+    //await clientPromise
+    const client = await clientPromise
+    const db = client.db('shoestore')
+
+    const results = await db.collection('products').find({}).toArray()
+    const resultstwo = await db.collection('accordianData').find({}).toArray()
+    const resultsthree = await db.collection('selectBarData').find({}).toArray()
+    const resultsfour = await db.collection('searchBarData').find({}).toArray()
+
+    const product = JSON.parse(JSON.stringify(results))
+    const accordian = JSON.parse(JSON.stringify(resultstwo))
+    const searchresults = JSON.parse(JSON.stringify(resultsfour))
+    const selectresults = JSON.parse(JSON.stringify(resultsthree))
+
+    return {
+      props:  {
+        allData :{product,accordian,searchresults,selectresults},
+        revalidate: 10,
+      }
     }
+    // res.status(200).send({product,accordian,searchresults,selectresults});
+
+  } catch (e) {
+    console.error(e)
   }
+  // return {
+  //   props:  {
+  //     allData,
+  //     revalidate: 10,
+  //   }
+  // }
 }
